fix protocol store tests
